Converted Lambda packager into a language-based extension point, other plugins can register different packaging.
