+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/upload/LambdaCreator.kt
@@ -1,12 +1,8 @@
 package software.aws.toolkits.jetbrains.services.lambda.upload
 
+import com.intellij.lang.Language
 import com.intellij.openapi.application.ApplicationManager
-import com.intellij.openapi.compiler.CompilerManager
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.roots.OrderRootType
-import com.intellij.openapi.roots.libraries.LibraryTablesRegistrar
-import com.intellij.openapi.vfs.VfsUtil
-import com.intellij.openapi.vfs.VirtualFile
 import software.amazon.awssdk.services.lambda.LambdaClient
 import software.amazon.awssdk.services.lambda.model.CreateFunctionRequest
 import software.amazon.awssdk.services.lambda.model.FunctionCode
@@ -15,19 +11,16 @@ import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
 import software.aws.toolkits.jetbrains.core.AwsClientManager
 import software.aws.toolkits.jetbrains.services.lambda.LambdaFunction
+import software.aws.toolkits.jetbrains.services.lambda.LambdaPackager
+import software.aws.toolkits.jetbrains.services.lambda.LambdaPackagerProvider
 import software.aws.toolkits.jetbrains.services.lambda.toDataClass
-import java.io.File
-import java.nio.file.Files
 import java.nio.file.Path
-import java.nio.file.Paths
-import java.util.stream.Collectors
-import java.util.zip.ZipEntry
-import java.util.zip.ZipOutputStream
 
 object LambdaCreatorFactory {
-    fun create(clientManager: AwsClientManager): LambdaCreator {
+    fun create(clientManager: AwsClientManager, language: Language): LambdaCreator {
+        val packager = LambdaPackagerProvider.getInstance(language)
         return LambdaCreator(
-                LambdaPackager(),
+                packager,
                 CodeUploader(clientManager.getClient()),
                 LambdaFunctionCreator(clientManager.getClient())
         )
@@ -40,7 +33,7 @@ class LambdaCreator(
     private val functionCreator: LambdaFunctionCreator
 ) {
     fun createLambda(functionDetails: FunctionUploadDetails, project: Project, onComplete: (LambdaFunction) -> Unit) {
-        packager.doPackaging(project) {
+        packager.createPackage(project) {
             uploader.upload(functionDetails, it) { key, version ->
                 functionCreator.create(functionDetails, key, version, onComplete)
             }
@@ -83,47 +76,5 @@ class CodeUploader(private val s3Client: S3Client) {
             val result = s3Client.putObject(por, code)
             onComplete(key, result.versionId())
         }
-    }
-}
-
-class LambdaPackager {
-    fun doPackaging(project: Project, onComplete: (Path) -> Unit) {
-        CompilerManager.getInstance(project).rebuild { aborted, errors, _, compileContext ->
-            if (!aborted && errors == 0) {
-                val classes = compileContext.projectCompileScope.affectedModules
-                        .map { compileContext.getModuleOutputDirectory(it) }
-                        .flatMap {
-                            val outputDir = it?.toPath()
-                            Files.walk(outputDir)
-                                    .filter { it.toString().toLowerCase().endsWith(".class") }
-                                    .map { Pair(outputDir?.relativize(it), it) }.collect(Collectors.toList<Pair<Path?, Path>>())
-                        }.filterNotNull()
-
-                val dependencies = LibraryTablesRegistrar.getInstance().getLibraryTable(project).libraries
-                        .flatMap { it.getFiles(OrderRootType.CLASSES).toList() }
-                        .map { VfsUtil.getVirtualFileForJar(it) }
-                        .map { it?.toPath() }.filterNotNull()
-
-                val zipFile = Files.createTempFile("function", ".zip")
-                val zip = ZipOutputStream(Files.newOutputStream(zipFile))
-
-                dependencies.forEach { addEntry("lib/${it.fileName}", it, zip) }
-                classes.forEach { addEntry(it.first.toString(), it.second, zip) }
-
-                zip.close()
-                onComplete(zipFile)
-            }
-        }
-    }
-
-    private fun addEntry(entryName: String, file: Path, zip: ZipOutputStream) {
-        zip.putNextEntry(ZipEntry(entryName))
-        val bytes = Files.readAllBytes(file)
-        zip.write(bytes, 0, bytes.size)
-        zip.closeEntry()
-    }
-
-    private fun VirtualFile.toPath(): Path {
-        return Paths.get(File(this.path).toURI())
     }
 }
