+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/upload/UploadLambdaFunction.kt
@@ -1,6 +1,5 @@
 package software.aws.toolkits.jetbrains.services.lambda.upload
 
-import com.intellij.lang.java.JavaLanguage
 import com.intellij.notification.Notification
 import com.intellij.notification.NotificationListener
 import com.intellij.notification.NotificationType
@@ -11,12 +10,13 @@ import com.intellij.openapi.actionSystem.LangDataKeys
 import com.intellij.openapi.fileEditor.FileEditorManager
 import software.amazon.awssdk.services.s3.model.Bucket
 import software.aws.toolkits.jetbrains.core.AwsClientManager
+import software.aws.toolkits.jetbrains.services.lambda.LambdaPackagerProvider
 import software.aws.toolkits.jetbrains.services.lambda.LambdaVirtualFile
 
 class UploadLambdaFunction : AnAction() {
 
     override fun update(e: AnActionEvent?) {
-        e?.presentation?.isEnabledAndVisible = e?.getData(LangDataKeys.PSI_FILE)?.language?.`is`(JavaLanguage.INSTANCE) == true
+        e?.presentation?.isEnabledAndVisible = e?.getData(LangDataKeys.PSI_FILE)?.language?.let { LambdaPackagerProvider.supportedLanguages().contains(it) } == true
     }
 
     override fun actionPerformed(event: AnActionEvent?) {
@@ -28,13 +28,8 @@ class UploadLambdaFunction : AnAction() {
             return
         }
 
-        if (!psi.language.`is`(JavaLanguage.INSTANCE)) {
-            handleError("Invalid language, only Java supported at present, language detected as '${psi.language}'")
-            return
-        }
-
         val uploadModal = UploadToLambdaModal(project, psi) { functionDetails ->
-            LambdaCreatorFactory.create(AwsClientManager.getInstance(project)).createLambda(functionDetails, project) {
+            LambdaCreatorFactory.create(AwsClientManager.getInstance(project), psi.language).createLambda(functionDetails, project) {
                 val notificationListener = NotificationListener { _, _ ->
                     val editorManager = FileEditorManager.getInstance(project)
                     val lambdaVirtualFile = LambdaVirtualFile(AwsClientManager.getInstance(project).getClient(), it)
